Moved non handler code to plugins/
