+++ plugins/database/ddbdoc-get-item.ts
@@ -1,5 +1,5 @@
 import {GetCommand} from "@aws-sdk/lib-dynamodb";
-import {ddbDocClient} from "./ddb-doc-client.mjs";
+import {ddbDocClient} from "./ddb-doc-client";
 
 export const getItem = async (tableName: string, primaryKey: string) => {
     try {
